refactor: restructure components by removing old implementations and consolidating exports
