+++ app/src/Components/Debugger/index.js
@@ -1,0 +1,2 @@
+import Debugger from "./Debugger";
+export default Debugger;
