+++ app/src/Components/DropArea/index.js
@@ -1,0 +1,2 @@
+import DropArea from "./DropArea";
+export default DropArea;
