+++ app/src/Components/SimulationTables/index.js
@@ -1,0 +1,2 @@
+import SimulationTables from "./SimulationTables";
+export default SimulationTables;
