+++ app/src/Components/index.js
@@ -1,0 +1,8 @@
+import Debugger from "./Debugger";
+import SimulationTables from "./SimulationTables";
+import DropArea from "./DropArea";
+export {
+    SimulationTables,
+    Debugger,
+    DropArea
+};
